chore(rds): move internal cluster interfaces and classes to own file
See https://github.com/aws/aws-cdk/pull/10130#pullrequestreview-482250360
for original motivation.

Moves `DatabaseClusterBaseProps`, `DatabaseClusterBase`, `DatabaseClusterNew`,
and `ImportedDatabaseCluster` from `cluster.ts` to
`private/cluster-internal.ts`.

Note that `DatabaseClusterBase` was previously publicly exported; however, that
change has not yet been release, so as long as this is merged before the next
release this won't be a breaking change.

Originally, I thought this refactoring was blocked by
https://github.com/aws/jsii/issues/1980
but thanks to a clarification, I realized I just needed to mark the private
classes as `@internal`.
